File validate data refresh , notification,UI changes , bug fixes

+++ FONTEND/openhack/src/app/bulkupload/bulkupload.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmComponent } from './../confirm/confirm.component';
 import { Filemetadata } from './../filemetadata';
 import { Component, OnInit, ViewChild, ChangeDetectorRef } from '@angular/core';
 import { AppService } from '../app.service';
@@ -6,6 +7,7 @@ import { UploadComponent } from './upload/upload.component';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-bulkupload',
@@ -17,6 +19,7 @@ export class BulkuploadComponent implements OnInit {
   displayedColumns: string[] = [
     'position',
     'fileName',
+    'createdDate',
     'validateStatus',
     'update',
     'delete'
@@ -28,19 +31,34 @@ export class BulkuploadComponent implements OnInit {
   constructor(
     private appService: AppService,
     private dialog: MatDialog,
-    private changeDetectorRefs: ChangeDetectorRef
+    private changeDetectorRefs: ChangeDetectorRef,
+    public snackBar: MatSnackBar
   ) {}
 
   ngOnInit() {
-    this.getValidatedFilesData();
+    this.getValidatedFilesData('INIT');
   }
 
-  getValidatedFilesData() {
+  getValidatedFilesData(type: string) {
     this.appService.getValidatedFilesData().subscribe(
       (response: Filemetadata[]) => {
         this.dataSource.data = response as Filemetadata[];
+        this.dataSource.sortingDataAccessor = (item, property) => {
+          switch (property) {
+            case 'createdDate':
+              return new Date(item.createdDate);
+            default:
+              return item[property];
+          }
+        };
         this.dataSource.sort = this.sort;
         this.dataSource.paginator = this.paginator;
+        if(type === 'REFRESH') {
+        this.snackBar.open( this.dataSource.data.length + ' files data is refreshed.', 'Close', {
+         duration: 2000,
+      });
+    }
+
         console.log('Data : ' + response.toString);
       },
       error => console.log('error : ' + error),
@@ -57,6 +75,7 @@ export class BulkuploadComponent implements OnInit {
     dialogConfig.autoFocus = true;
     dialogConfig.minHeight = '500px';
     dialogConfig.minWidth = '500px';
+    dialogConfig.panelClass = 'mat-dialog-override';
     const dialogRef = this.dialog.open(UploadComponent, dialogConfig);
 
     dialogRef.afterClosed().subscribe(data => {
@@ -66,12 +85,34 @@ export class BulkuploadComponent implements OnInit {
           this.dataSource.data.push(item);
         }
         this.dataSource.data = [...this.dataSource.data];
+        this.snackBar.open(data.length + ' files uploaded.', 'Close', {
+         duration: 2000,
+      });
       }
     });
   }
 
-  deleteValidateFielData(validateFileObj) {
-    this.appService.deleteValidateFielData(validateFileObj).subscribe(
+  getUpdateDataById(currentElement) {
+    const validateId = currentElement.id;
+    let responseData: any  = {};
+    this.appService.getUpdateDataById(validateId).subscribe(
+      data => {
+        if ( data ) {
+          responseData = data;
+          this.dataSource.data[this.getIndexfromdataSource(validateId)].validateStatus = responseData.validateStatus;
+          this.dataSource.data = [...this.dataSource.data];
+          this.snackBar.open(currentElement.fileName + ' data refreshed.', 'Close', {
+         duration: 2000,
+      });
+        }
+        console.log('Id data' + data);
+      },
+      error => console.log('Error in getting data from ID')
+    );
+  }
+
+  deleteValidateFileData(validateFileObj) {
+    this.appService.deleteValidateFileData(validateFileObj).subscribe(
       response => {
         let tempList = [];
         tempList = this.dataSource.data.splice(
@@ -79,6 +120,9 @@ export class BulkuploadComponent implements OnInit {
           1
         );
         this.dataSource.data = [...this.dataSource.data];
+        this.snackBar.open(validateFileObj.fileName + ' is deleted.', 'Close', {
+         duration: 2000,
+      });
         console.log(response);
       },
       error => console.log(error),
@@ -89,13 +133,27 @@ export class BulkuploadComponent implements OnInit {
   }
 
   deleteAllValidateFielData() {
-    this.appService.deleteAllValidateFielData().subscribe(
+        const dialogRef = this.dialog.open(ConfirmComponent, {
+      width: '350px',
+      data: 'Do you confirm the deletion?'
+    });
+        dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        console.log('Yes clicked');
+        this.appService.deleteAllValidateFielData().subscribe(
       response => {
         console.log('complete data deleted');
+        const dataSourceLength = this.dataSource.data.length;
         this.dataSource.data = [];
+        this.snackBar.open(dataSourceLength + ' items deleted.', 'Close', {
+         duration: 2000,
+      });
       },
       error => console.log('error in data deletion')
     );
+      }
+    });
+
   }
 
   getIndexfromdataSource(validateFileId) {
